Updated front end
front end updates so that we have React front end that is compatible with .NET and Docker container, we will need to migrate the previous front end code into this.

When debugging, the JavaScript client runs on the host machine, but the ASP.NET Core server code runs in the container. When published, the proxy is not run, and the client code runs on the same server as the ASP.NET Core code